requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.28.1
idna==3.7
macholib==1.16.3
MouseInfo==0.1.3
moviepy==1.0.3
openai==1.66.3
packaging==24.2
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyMsgBox==1.0.9
pyobjc-core==10.1; sys_platform == 'darwin'
pyobjc-framework-Cocoa==10.1; sys_platform == 'darwin'
pyobjc-framework-Quartz==10.1; sys_platform == 'darwin'
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.0.7
requests==2.32.0
setuptools==75.3.0
sniffio==1.3.0
tqdm==4.66.4
ttkbootstrap==1.10.1
typing_extensions==4.12.2
urllib3==2.2.2
google-genai==1.5.0
